linkedin-automation/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0
pillow>=10.3.0
sqlalchemy>=2.0.25
aiohttp>=3.9.1
oagi>=0.12.0
pyperclip>=1.8.2

